+++ lib/src/canvas_context/diagram_editor_context.dart
@@ -13,12 +13,18 @@ class DiagramEditorContext {
   final CanvasModel _canvasModel;
   final CanvasState _canvasState;
 
+  /// Set of policies where all the diagram customization is defined.
   final PolicySet policySet;
 
+  /// Canvas model containing all components and links with all the functions.
   CanvasModel get canvasModel => _canvasModel;
 
+  /// Canvas state containing for example canvas position and scale.
   CanvasState get canvasState => _canvasState;
 
+  /// [DiagramEditorContext] is taken as parameter by [DiagramEditor] widget.
+  ///
+  /// Its not generated automatically because you want to use it to copy model or state to another [DiagramEditor].
   DiagramEditorContext({
     @required this.policySet,
   })  : this._canvasModel = CanvasModel(policySet),
@@ -26,7 +32,9 @@ class DiagramEditorContext {
     policySet.initializePolicy(_getReader(), _getWriter());
   }
 
-  /// Some policies are used in CanvasModel, so they will be shared as well.
+  /// Allows you to create [DiagramEditorContext] with shared model from another [DiagramEditorContext].
+  ///
+  /// Warning: [LinkAttachmentPolicy] is used in CanvasModel, so this policy will be shared as well, even if you put new one to [PolicySet].
   DiagramEditorContext.withSharedModel(
     DiagramEditorContext oldContext, {
     @required this.policySet,
@@ -35,6 +43,7 @@ class DiagramEditorContext {
     policySet.initializePolicy(_getReader(), _getWriter());
   }
 
+  /// Allows you to create [DiagramEditorContext] with shared state (eg. canvas position and scale) from another [DiagramEditorContext].
   DiagramEditorContext.withSharedState(
     DiagramEditorContext oldContext, {
     @required this.policySet,
@@ -43,7 +52,9 @@ class DiagramEditorContext {
     policySet.initializePolicy(_getReader(), _getWriter());
   }
 
-  /// Some policies are used in CanvasModel, so they will be shared as well.
+  /// Allows you to create [DiagramEditorContext] with shared model and state from another [DiagramEditorContext].
+  ///
+  /// Warning: [LinkAttachmentPolicy] is used in CanvasModel, so this policy will be shared as well, even if you put new one to [PolicySet].
   DiagramEditorContext.withSharedModelAndState(
     DiagramEditorContext oldContext, {
     @required this.policySet,