+++ lib/src/canvas_context/model/connection.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
 
 abstract class Connection {
+  /// Id of this connection. It corresponds to link id.
   final String connectionId;
+
+  /// Id of a component to which is the component connected.
   final String otherComponentId;
 
+  /// Abstract class that represents connection of a component.
   Connection({
     @required this.connectionId,
     @required this.otherComponentId,
@@ -15,6 +19,7 @@ abstract class Connection {
 }
 
 class ConnectionOut extends Connection {
+  /// Connection type that is saved to source component [connection].
   ConnectionOut({
     @required connectionId,
     @required otherComponentId,
@@ -25,6 +30,7 @@ class ConnectionOut extends Connection {
 }
 
 class ConnectionIn extends Connection {
+  /// Connection type that is saved to target component [connection].
   ConnectionIn({
     @required connectionId,
     @required otherComponentId,