+++ CHANGELOG.md
@@ -1,7 +1,25 @@
 
+# 0.0.9
+
+* Add documentation comments to reader/writer
+* Move custom components widgets from Canvas to Component
+
+# 0.0.8
+
+* Add some more documentation comments.
+* Fix documentation comments.
+
+# 0.0.7
+
+* Add more documentation comments.
+* Fix readme.
+
 # 0.0.6
 
-* Update example and add comments.
+* Add documentation comments.
+* Update the example and add comments to it.
+* Update readme file.
+* Add some asserts ("the model must contain this id").
 
 # 0.0.5